set ChartView ChartSeries in DB

+++ App_Code/SysServices/Designers/_Dashboard.cs
@@ -20,7 +20,6 @@ public class _Dashboard : System.Web.Services.WebService
         try
         {
 
-
             //GET JSON
             int id = Convert.ToInt32(HttpContext.Current.Request.Form["id"]);
             JavaScriptSerializer js = new JavaScriptSerializer();
@@ -46,26 +45,17 @@ public class _Dashboard : System.Web.Services.WebService
             //CREATE PAGE ACCESS
             CreatePageAccess(design.DashboardID, design.AccessID, design.Version);
 
-
+            //CREATE PAGE NAVIGATIONS
             int Nav = 1000;
+            for (int i = 0; i < 5; i++) CreatePageNavigations(design.DashboardID, Nav++, design.Version);
 
-            for (int i = 0; i < 5; i++)
-            {
-                //CREATE PAGE NAVIGATIONS
-                CreatePageNavigations(design.DashboardID, Nav++, design.Version);
-            }
-
-
-
+            //CREATE ROWS
             for (int i = 0; i < design.RowBoxs.Length; i++)
             {
 
                 //INTERNALID GENERATOR
                 design.RowBoxs[i].InternalID = design.DashboardID.ToString() + design.RowBoxs[i].RowIndex.ToString();
-
                 RowBox row = design.RowBoxs[i];
-
-                //CREATE ROWS
                 CreatePageElement(row.InternalID, design.DashboardID, "0", null, "Row", null, false, false, null, false, design.Version);
 
 
@@ -99,19 +89,18 @@ public class _Dashboard : System.Web.Services.WebService
                         //CREATE ACTIVITY CONTEXT
                         CreateActivityContext(activityID, chart.Name, chart.Text, "ChartView", null, design.Version, null);
 
+                        //CREATE ACTIVITY BOX
+                        CreatePageElement(chart.InternalID + "0", design.DashboardID, chart.InternalID, activityID.ToString(), "ActivityBox", "solid", false, false, null, false, design.Version);
+
                         //CREATE ACCESS ACTIVITIES
                         CreateAccessActivities(design.AccessID, activityID, design.Version);
 
-
                         //CHART VIEWS
-                        CreateChartViews(activityID, activityID, "column", "", "", "", "", "", "", "", "", "", "", "", "");
+                        CreateChartViews(activityID, activityID, chart.Type, chart.Text, "", chart.CategoryLabel, chart.ValueLabel, chart.CategoryName, chart.CategoryExpression, "simple", "", "", 0, 0, "");
 
-                        //createchart()
-
-                        //    forechar(chart.Series)
-
-                        //CREATE ACTIVITY BOX
-                        CreatePageElement(chart.InternalID + "0", design.DashboardID, chart.InternalID, activityID.ToString(), "ActivityBox", "solid", false, false, null, false, design.Version);
+                        //CHART Series
+                        foreach (var series in chart.Series)
+                            CreateChartSeries(Int32.Parse((activityID.ToString()) + "0"), activityID, series.Name, series.Text, series.DataExpression, 0, series.PlotType, series.StyleColor, 1, chart.Version, "", "", null, null);
 
                         activityID++;
                     }
@@ -138,20 +127,37 @@ public class _Dashboard : System.Web.Services.WebService
 
     public static void ClearTables(int activityID, Array charts)
     {
-        SqlDataProvider.ExecuteNoneQuery(string.Format("DELETE FROM Sys_Gui_Pages WHERE PageID=1000010;" +
-            "DELETE FROM[FS_Software].[dbo].[Sys_Gui_PageLayouts] WHERE PageID = 1000010;" +
-            "DELETE FROM Sys_Dev_Activities WHERE ActivityID in (select ActivityID FROM[Sys_Activities] WHERE ModuleID = 15010);" +
-            "DELETE FROM Sys_Gui_ActivityContexts WHERE ActivityID in (select ActivityID FROM[Sys_Activities] WHERE ModuleID = 15010);" +
-            "DELETE FROM[Sys_Activities] WHERE ModuleID = 15010;" +
-            "DELETE FROM[FS_Software].[dbo].Sys_Gui_Page_Contexts WHERE PageID=1000010;" +
-            "DELETE FROM Sys_Gui_Page_Accesses WHERE PageID=1000010;" +
-            "DELETE FROM Sys_Access_Activities WHERE AccessID=150101;" +
-            "DELETE FROM Sys_Gui_Page_Navigations WHERE PageID=1000010;"));
-        for (int i = 0; i < charts.Length; i++)
-        {
-            SqlDataProvider.ExecuteNoneQuery(string.Format("DELETE FROM Sys_Gui_ChartViews WHERE ChartID=1501050") );
+        SqlDataProvider.ExecuteNoneQuery(string.Format("DELETE FROM Sys_Gui_ChartSeries WHERE ChartID = 1501050;" +
+            "DELETE FROM Sys_Gui_ChartSeries WHERE ChartID = 1501051;" +
+            "DELETE FROM Sys_Gui_ChartSeries WHERE ChartID = 1501052;" +
+            "DELETE FROM Sys_Gui_ChartSeries WHERE ChartID = 1501053;" +
+            "DELETE FROM Sys_Gui_ChartSeries WHERE ChartID = 1501054;" +
+             "DELETE FROM Sys_Gui_ChartSeries WHERE ChartID = 1501055;" +
+              "DELETE FROM Sys_Gui_ChartSeries WHERE ChartID = 1501056;" +
+               "DELETE FROM Sys_Gui_ChartSeries WHERE ChartID = 1501057;" +
+                "DELETE FROM Sys_Gui_ChartSeries WHERE ChartID = 1501058;" +
 
-        }
+            "DELETE FROM Sys_Gui_ChartViews WHERE ChartID = 1501050;" +
+            "DELETE FROM Sys_Gui_ChartViews WHERE ChartID = 1501051;" +
+            "DELETE FROM Sys_Gui_ChartViews WHERE ChartID = 1501052;" +
+            "DELETE FROM Sys_Gui_ChartViews WHERE ChartID = 1501053;" +
+            "DELETE FROM Sys_Gui_ChartViews WHERE ChartID = 1501054;" +
+             "DELETE FROM Sys_Gui_ChartViews WHERE ChartID = 1501055;" +
+              "DELETE FROM Sys_Gui_ChartViews WHERE ChartID = 1501056;" +
+               "DELETE FROM Sys_Gui_ChartViews WHERE ChartID = 1501057;" +
+                "DELETE FROM Sys_Gui_ChartViews WHERE ChartID = 1501058;" +
+            "DELETE FROM Sys_Gui_PageLayouts WHERE PageID = 1000010;" +
+            "DELETE FROM Sys_Dev_Activities WHERE ActivityID in (select ActivityID FROM[Sys_Activities] WHERE ModuleID = 15010);" +
+            "DELETE FROM Sys_Gui_Page_Accesses WHERE PageID = 1000010;" +
+            "DELETE FROM Sys_Access_Activities WHERE AccessID = 150101;" +
+            "DELETE FROM dbo.Sys_Access_Activities WHERE ActivityID in (select ActivityID FROM[Sys_Activities] WHERE ModuleID = 15010);" +
+            "DELETE FROM Sys_Gui_ActivityContexts WHERE ActivityID in (select ActivityID FROM[Sys_Activities] WHERE ModuleID = 15010) ;" +
+            "DELETE FROM Sys_Gui_Page_Contexts WHERE PageID = 1000010;" +
+            "DELETE FROM Sys_Gui_Page_Navigations WHERE PageID = 1000010;" +
+            "DELETE FROM Sys_Activities WHERE ModuleID = 15010;" +
+            "DELETE FROM Sys_Gui_Pages WHERE PageID = 1000010;"
+                     ));
+
     }
 
     public static void CreatePage(int pageId, string PageTemplateID, string name, string label, string type, bool headerVisible, string version, string desc)
@@ -202,10 +208,17 @@ public class _Dashboard : System.Web.Services.WebService
         SqlDataProvider.ExecuteNoneQuery(string.Format("INSERT INTO Sys_Gui_Page_Navigations(pageID,NavigationID,version) VALUES({0},{1},N'{2}')", pageID, NavigationID, version));
     }
 
-    public static void CreateChartViews(int ChartID, int ActivityContextID, string ChartType, string Title, string SubTitle, string CategoryLabel, string ValueLabel, string CategoryName, string CategoryExpression, string SeriesType, string GroupingExpression, string TimePrioiedType, string TimePrioiedStart, string TimePrioiedDuration, string TimePrioiedSeriesFormat)
+    public static void CreateChartViews(int ChartID, int ActivityContextID, string ChartType, string Title, string SubTitle, string CategoryLabel, string ValueLabel, string CategoryName, string CategoryExpression, string SeriesType, string GroupingExpression, string TimePrioiedType, int TimePrioiedStart, int TimePrioiedDuration, string TimePrioiedSeriesFormat)
     {
-        SqlDataProvider.ExecuteNoneQuery(string.Format("INSERT INTO Sys_Gui_ChartViews(ChartID,ActivityContextID,ChartType,Title,SubTitle,CategoryLabel,ValueLabel,CategoryName,CategoryExpression,SeriesType,GroupingExpression,TimePrioiedType,TimePrioiedStart,TimePrioiedDuration,TimePrioiedSeriesFormat) VALUES({0},{1},N'{2}',N'{3}',N'{4}',N'{5}',N'{6}',N'{7}',N'{8}',N'{9}',N'{10}',N'{11}',N'{12}',N'{13}',N'{14}')", ChartID, ActivityContextID, ChartType, Title, SubTitle, CategoryLabel, ValueLabel, CategoryName, CategoryExpression, SeriesType, GroupingExpression, TimePrioiedType, TimePrioiedStart, TimePrioiedDuration, TimePrioiedSeriesFormat));
+        SqlDataProvider.ExecuteNoneQuery(string.Format("INSERT INTO Sys_Gui_ChartViews(ChartID,ActivityContextID,ChartType,Title,SubTitle,CategoryLabel,ValueLabel,CategoryName,CategoryExpression,SeriesType,GroupingExpression,TimePrioiedType,TimePrioiedStart,TimePrioiedDuration,TimePrioiedSeriesFormat) VALUES({0},{1},N'{2}',N'{3}',N'{4}',N'{5}',N'{6}',N'{7}',N'{8}',N'{9}',N'{10}',N'{11}',{12},{13},N'{14}')", ChartID, ActivityContextID, ChartType, Title, SubTitle, CategoryLabel, ValueLabel, CategoryName, CategoryExpression, SeriesType, GroupingExpression, TimePrioiedType, TimePrioiedStart, TimePrioiedDuration, TimePrioiedSeriesFormat));
     }
+
+    public static void CreateChartSeries(int ChartSeriesID, int ChartID, string Name, string Label, string DataExpression, int SeriesIndex, string PlotType, string StyleColor, int Enabled, string Version, string Description, string ActionOnClick, string LanchedContextID, string LanchedPageID)
+    {
+        SqlDataProvider.ExecuteNoneQuery(string.Format("INSERT INTO Sys_Gui_ChartSeries(ChartSeriesID,ChartID,Name,Label,DataExpression,SeriesIndex,PlotType,StyleColor,Enabled,Version,Description,ActionOnClick,LanchedContextID,LanchedPageID) VALUES({0},{1},N'{2}',N'{3}',N'{4}',N'{5}',N'{6}',N'{7}',{8},N'{9}',N'{10}',N'{11}',{12},{13})", ChartSeriesID, ChartID, Name, Label, DataExpression, SeriesIndex, PlotType, StyleColor, Enabled, Version, Description, ActionOnClick, (LanchedPageID == null ? "NULL" : LanchedContextID), (LanchedPageID == null ? "NULL" : LanchedPageID)));
+    }
+
+
 
 }
